+++ src/app/core/components/media/media-gallery-dialog/media-gallery-dialog.component.html
@@ -3,7 +3,7 @@
         <div style="width:max-content;">
             {{ gallery.date}}
             <div fxLayout="row" fxLayoutGap="0.5rem">
-                <media-thumbnails [gallery]="gallery"></media-thumbnails>
+                <media-thumbnails [gallery]="gallery.images"></media-thumbnails>
             </div>
         </div>
     </ng-container>
