--- src/app/core/components/media/media-gallery-dialog/media-gallery-dialog.component.scss
@@ -1,23 +1,0 @@
-.mediaGalleryImages{
-    width: 11rem;
-    min-height: 8rem; 
-    background-size: cover;
-    background-repeat: no-repeat;
-    border-radius:0.2rem;
-    position: relative;
-}
-.imagesTime{
-    position: absolute;
-    margin: 0.5rem;
-    color: #fff;
-    font-weight: bold;
-    left: 0;
-    top: 0;
-}
-.mediaGalleryImages:before {
-    display: block;
-    content: "";
-    width: 100%;
-    height: 100%;
-    background: linear-gradient(135deg, rgba(0,0,0,0.7), rgba(255,255,255,0.1));
-}
