--- src/app/core/components/media/media-gallery-dialog/media-gallery-dialog.module.ts
@@ -1,22 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { MaterialModule } from 'src/app/material-module';
-import { FlexLayoutModule } from '@angular/flex-layout';
-import { MediaGalleryDialogComponent } from './media-gallery-dialog.component';
-import { MediaThumbnailsModule } from '../media-thumbnails/media-thumbnails.module';
-
-
-
-@NgModule({
-  declarations: [
-    MediaGalleryDialogComponent
-  ],
-  imports: [
-    CommonModule,
-    MaterialModule,
-    FlexLayoutModule,
-    MediaThumbnailsModule
-  ],
-  exports: [MediaGalleryDialogComponent]
-})
-export class MediaGalleryDialogModule { }
