+++ src/app/core/components/media/media-gallery/media-gallery.component.html
@@ -1,11 +1,6 @@
 
 <div fxLayout="row" class="mediaGalleryContainer"> 
-    <!-- <ng-container *ngFor="let image of this.images.slice(0,3)">
-        <div class="mediaGalleryImages" [style.backgroundImage]="'url('+ image.thumbnail +')'">
-            <span class="imagesTime">{{image.time}}</span>
-        </div>
-    </ng-container> -->
-    <media-thumbnails></media-thumbnails>
+    <media-thumbnails [gallery]="this.images.slice(0,3)"></media-thumbnails>
 </div>
 <div fxLayout="row" fxLayoutAlign="end center">
     <button type='button' color="primary" mat-mini-fab>
