+++ src/app/core/components/media/media-gallery/media-gallery.component.scss
@@ -1,7 +1,6 @@
 .mediaGalleryContainer{
     overflow-x:scroll;
     scroll-behavior: smooth;
-    justify-content: space-between;
     &::-webkit-scrollbar {
         background-color: #fff;
         width: 0.8rem;
@@ -23,27 +22,4 @@
     &::-webkit-scrollbar-button {
         display:none;
     }
-}
-.mediaGalleryImages{
-    min-width: calc(100%/3 - 0.33rem); 
-    min-height: 8rem; 
-    background-size: cover;
-    background-repeat: no-repeat;
-    border-radius:0.2rem;
-    position: relative;
-}
-.imagesTime{
-    position: absolute;
-    margin: 0.5rem;
-    color: #fff;
-    font-weight: bold;
-    left: 0;
-    top: 0;
-}
-.mediaGalleryImages:before {
-    display: block;
-    content: "";
-    width: 100%;
-    height: 100%;
-    background: linear-gradient(135deg, rgba(0,0,0,0.7), rgba(255,255,255,0.1));
 }
