+++ src/app/core/components/media/media-gallery/media-gallery.module.ts
@@ -4,9 +4,6 @@ import { MediaGalleryComponent } from './media-gallery.component';
 import { MaterialModule } from 'src/app/material-module';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { MediaThumbnailsModule } from '../media-thumbnails/media-thumbnails.module';
-
-
-
 @NgModule({
   declarations: [
     MediaGalleryComponent
