+++ src/app/core/components/media/media-thumbnails/media-thumbnails.component.ts
@@ -1,12 +1,17 @@
-import { Component, Input } from '@angular/core';
-
+import { Component, Input, OnInit } from '@angular/core';
+import { Gallery, Image } from 'src/app/core/models/media';
 @Component({
     // tslint:disable-next-line: component-selector
     selector: 'media-thumbnails',
     templateUrl: './media-thumbnails.component.html',
     styleUrls: ['./media-thumbnails.component.scss']
 })
-export class MediaThumbnailsComponent {
-    @Input() gallery!:any;
+export class MediaThumbnailsComponent implements OnInit{
+    @Input() gallery!:Image[];
+
     constructor () {}
+
+    ngOnInit(): void {
+       
+    }
 }
