+++ src/app/core/components/media/media-thumbnails/media-thumbnails.module.ts
