+++ src/app/core/components/thumbnail-slider/thumbnail-slider.module.ts
@@ -1,8 +1,8 @@
+import { MediaGalleryModule } from './../media/media-gallery/media-gallery.module';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { NgxGalleryModule } from '@animalaidunlimited/ngx-gallery-aau';
 import { MaterialModule } from '../../../material-module';
-import { MediaGalleryModule } from '../media/media-gallery/media-gallery.module';
 import { ThumbnailSliderComponent } from '../thumbnail-slider/thumbnail-slider.component';
 
 @NgModule({
